config: change UID fields from `u16` to `u32`
Changes `min_uid` and `max_uid` from `u16` to `u32` in order to support
the whole Unix UID range (which is up to 4,294,967,295).

+++ src/config/env.rs
@@ -107,7 +107,7 @@ pub fn apply_env_vars(config: &mut Config) {
   }
 
   if let Ok(value) = env::var("TUIGREET_USER_MENU_MIN_UID") {
-    if let Ok(uid) = value.parse::<u16>() {
+    if let Ok(uid) = value.parse::<u32>() {
       config.user_menu.min_uid = uid;
     } else {
       tracing::warn!(
@@ -118,7 +118,7 @@ pub fn apply_env_vars(config: &mut Config) {
   }
 
   if let Ok(value) = env::var("TUIGREET_USER_MENU_MAX_UID") {
-    if let Ok(uid) = value.parse::<u16>() {
+    if let Ok(uid) = value.parse::<u32>() {
       config.user_menu.max_uid = uid;
     } else {
       tracing::warn!(
